Implement Update User

+++ src/main/java/rookiesspring/service/UserService.java
@@ -5,14 +5,19 @@
 package rookiesspring.service;
 
 import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rookiesspring.dto.UserDTO;
 import rookiesspring.dto.response.UserResponseDTO;
 import rookiesspring.dto.response.custom.UserResponseDTOShort;
-import rookiesspring.exception.ResourceNotFoundException;
+import rookiesspring.dto.update.UserUpdateDTO;
 import rookiesspring.mapper.UserMapper;
+import rookiesspring.mapper.UserUpdateMapper;
+import rookiesspring.model.Address;
 import rookiesspring.model.User;
+import rookiesspring.model.UserDetail;
 import rookiesspring.repository.UserRepository;
 import rookiesspring.service.interfaces.UserServiceInterface;
 
@@ -23,35 +28,39 @@ import rookiesspring.service.interfaces.UserServiceInterface;
  */
 @Service
 public class UserService implements UserServiceInterface {
-
+    
     private UserRepository repository;
     private UserMapper mapper;
-
+    @Autowired
+    private UserUpdateMapper updateMapper;
+    
     public UserService(UserRepository repository, UserMapper mapper) {
         this.repository = repository;
         this.mapper = mapper;
     }
-
-    public List<UserResponseDTOShort> findAll() {
-        return repository.findAllProjectedBy();
-    }
-
+    
     public UserResponseDTOShort findById(Long userId) {
-        return repository.findProjectedById(userId).orElseThrow(() -> new ResourceNotFoundException());
+        return repository.findProjectedById(userId).orElseThrow(() -> new EntityNotFoundException());
     }
-
+    
     public List<UserResponseDTOShort> findAllByUsername(String username) {
+        if (username == null) {
+            username = "";
+        }
         return repository.findAllProjectedByUsernameContainsIgnoreCase(username);
     }
-
-    public List<UserResponseDTO> findAllFull() {
-        return mapper.ToResponseDTOList(repository.findAll());
+    
+    public List<UserResponseDTO> findAllFull(String username) {
+        if (username == null) {
+            username = "";
+        }
+        return mapper.ToResponseDTOList(repository.findAll(username));
     }
-
+    
     public UserResponseDTO findByIdFull(Long userId) {
-        return mapper.ToResponseDTO(repository.findById(userId).orElseThrow(() -> new ResourceNotFoundException()));
+        return mapper.ToResponseDTO(repository.findById(userId).orElseThrow(() -> new EntityNotFoundException()));
     }
-
+    
     public UserResponseDTO save(UserDTO newUser) {
         User u = mapper.toEntity(newUser);
         try {
@@ -60,17 +69,46 @@ public class UserService implements UserServiceInterface {
             throw new EntityExistsException();
         }
     }
+// doing
 
-    public UserResponseDTO updateOne(User oldUser) {
-        return mapper.ToResponseDTO(repository.save(oldUser));
+    public UserResponseDTOShort updateOne(UserUpdateDTO user_dto) {
+        if (checkExist(user_dto.id())) {
+            User u = repository.getReferenceById(user_dto.id());
+            if (user_dto.email() != null) {
+                if (checkExistEmail(user_dto.email())) {
+                    throw new EntityExistsException("Email has already Exists");
+                }
+                
+            }
+            Address address = new Address();
+            updateMapper.updateUserAddressFromDto(user_dto, address);
+            UserDetail detail = new UserDetail();
+            updateMapper.updateUserDetailFromDto(user_dto, detail);
+            updateMapper.updateUserFromDto(user_dto, u);
+            detail.setAddress(address);
+            u.setUser_detail(detail);
+            
+            repository.save(u);
+            return mapper.ToResponseDTOShort(u);
+        } else {
+            throw new EntityNotFoundException("No value present");
+        }
     }
-
+    
     public void deleteById(long id) {
-        repository.deleteById(id);
+        if (checkExist(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException();
+        }
     }
-
+    
     @Override
     public boolean checkExist(long id) {
         return repository.existsById(id);
+    }
+    
+    public boolean checkExistEmail(String email) {
+        return repository.existsByEmail(email);
     }
 }
